Fixed bugs. Tangent application is still buggy, causing incorrect surface mesh to be produced.

+++ src/brep_builder.hxx
@@ -20,8 +20,8 @@ class CylindricalTool
 {
 public:
 
-    int height;
-    int radius;
+    double height;
+    double radius;
 };
 
 /*
